[Fluid] Move test and extend from abstract test
